+++ config/default.local.js
@@ -1,6 +1,0 @@
-'use strict';
-
-module.exports = {
-	port: 8400,
-	bind: '0.0.0.0',
-};