+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## v0.1.2 (2018-03-28)
+
+Removed configuration files.
+Added configuration via service exports.
+Added support for multiple services.
+Improved documentation.
+
 ## v0.1.1 (2018-03-27)
 
-Added CORS.
+Added CORS support.